Clean up titles and parsing a bit
Now + decodes to space in the triples file, and you can set the title of
the bundle to 'plic' there with 'dercuano bundle plic'.  And the title
of each note has '⁑ Dercuano' (or whatever the bundle title is) appended
to it, so that the HTML titles will be globally unique.  This involved a
bit of Unicode fixing.

Next up: linking the note from the index page, linking it from its
category pages, linking the category pages from the note page, linking
the category pages from the index page.