Shave a third off minimal runtime by caching dates
Dercuano was taking about 3.7 seconds to run on my laptop even when it
didn’t rerender any notes.  Now it takes 2.5 seconds instead, about one
third less.  That’s because now it’s not iterating over all the tuples
in the triple store every time it needs the date for a post to sort a
category.  (There’s an obvious solution to the tuple-iteration problem
which I ought to implement...)