Simplify Dercuano code somewhat
There was some extra caching and indirection in there that was no longer
needed, if it ever was.  (Caching the list of categories for each note
was actually needed at one point, but now the triple store is fast
enough that it doesn’t help.  So I added the caching, sped up the triple
store, and removed the caching; each step was an incremental
improvement, but together they sound stupid.  My consolation is that
adding the caching was three lines of code, and removing it was easy).

Also there were some cases where Note objects were getting instantiated
redundantly, even though the Bundle instantiates all the Notes at
startup nowadays, so the code to do that could be removed along with all
of its safeguards, replaced with a simple dictionary lookup.

I also reordered a bit of code to make it easier to read.

All this actually sped the build process up from about 18.1 seconds to
about 17.6 seconds, as well as shortening the code by about 23
lines (out of 364, according to cloc, so this is about a 7%
improvement in complexity — nothing to sneeze at!).