Allow repeated subjects in triples with indent
I was getting sick of typing *and reading* the same note name over and
over again on different lines, so this patch allows the implicit
repetition of the previous subject name in the triples file, signaled by
indentation.  Thus the sequence of three lines

    transmission-line-diode-computation written 2016-07-30
        updated 2016-07-30
        concerns electronics physical-computation pricing

is equivalent to the still-valid sequence of four lines

    transmission-line-diode-computation written 2016-07-30
    transmission-line-diode-computation updated 2016-07-30
    transmission-line-diode-computation concerns electronics
    transmission-line-diode-computation concerns physical-computation pricing

but much easier to read.

I changed the triples file accordingly, and I’ve verified that the
rendered output is byte-for-byte identical with `diff -urN`.  (On the
first comparison, I’d made a mistake in editing triples and had to fix
it.)