Tie some things together
Added more notes on lithium fission, and added `big-if-true` as a sort
of attempt to organize some of the big recurring ideas.  Also I updated
the intro disclaimer to add that even where correct, Dercuano may not be
original.

I also fixed a bunch of metadata things I noticed as I was poking around
writing `big-if-true`.