Fix rendering of `implementation–separation`
In particular, the Unicode line-drawing characters weren’t meeting as
they should in `<pre>`, with gaps between lines of text, so I tightened
up the line-height a bit more to 0.9, even though that makes code look
worse.  This made them (usually) overlap, so I turned off translucency
for `<pre>` text so that wouldn’t make black spots.  I’m not super happy
with this result.

Also, this note has an `<h3>` that begins a `<blockquote>` (being used
as an aside), and the extra margin above the `<h3>` looked out of place,
so I extended the `h1:first-child` layout rule to all the header types.

And I found out that Markdown, at least markdown-python, doesn’t let you
put a link target definition inside a `<blockquote>` block, so I fixed
that.