Write about the 8080 and sandwich panels
Over a month ago I printed out the 8080 instruction set and started
drawing opcode tables on paper, and I ran across the paper today, so I
put together the HTML version of the 8080 opcode table that Intel should
have published in the first place.  Bleh.

Also I had a bunch of pending stuff about sandwich panels, some of which
got lost when my netbook broke, so I rewrote a bunch of it here, but
with less information about foaming plastics.