Fix more jankiness in PDF generation
Now it starts the PDF file with the main table of contents, although
the links in it are inoperative.

Also, this no longer renders the <title> as actual text (thus
postponing the tofu from a ⁑ in the first line until somewhere later
on) and does treat a <div> as a block-level element.  And I
reorganized and updated the to-do list.