Fix ElementTidy parsing of non-XML
Some 3% of Dercuano's notes were failing to parse with ElementTree, so
I tried ElementTidy, which wasn't doing what I hoped for --- it was
producing mojibake and not recognizing any tags.  Here are some hacks
to fix that.  Also there is a hack in here to avoid displaying scripts
and CSS styles as body text.