Change PDF generation to use less stacks
Almost, but not quite, "refactor".

The janky PDF generation code was using a separate stack for restoring
font sizes when it exited elements, which was goofy because it had to
find out *when* to restore the font by looking at the element stack.
Also it had a totally goofy way of handling element tails which
necessitated more complexity in setting fonts.

This replaces all of that with a "current_style" dictionary whose
values get restored from values stored on the normal stack.  And now
link destinations are stored in there along with fonts.

This sets the stage for implementing `<pre>`.

This is noticeably slower now; 6.5 minutes instead of 4.5 to generate
the whole Dercuano PDF.  Presumably that's because it's setting the
font redundantly on every word.

I drive-by fixed the computations of bounding boxes for links;
previously the first line of the link had its baseline pushed down by
0.1 ems, while subsequent lines wouldn't.  This was a bug, so now it's
always pushed down.  That's why this is not a refactoring.