Fix links and bottom margin in PDF generation
The PDF generator had a very vague idea about where its Y position on
the page was, and this resulted in links not actually being in the
right place on the page.  Some more quick hacks to PDF textobject
generation to generate one textobject per line, and now links work.
However, the PDF is now more than 1MB larger.

Also, I tweaked the link boxes to have a little padding on the left
side, and bumped the version number.