Fix more PDF generation problems
Specifically, this solves the problem in `escheme` where
character-level markup inside a `<pre>` would produce spurious
newlines, and it also adds some margins above and below the `<pre>`;
and it fixes the problem with URL-encoded links to `lua-#-operator`
etc.  Also I updated the comments and added a little padding inside
the right side of link rectangles and updated the to-do list.