Fix PDF line spacing with subscripts/superscripts
There was a bug in the PDF renderer when superscripts or especially
subscripts were present; this made lines with lots of subscripts
totally unreadable because several lines would overlap.  Now this bug
is fixed in most cases, and less severe in the others.

Also fixed minor formatting and spelling problems in
`complex-linear-regression` and `dercuano-stylesheet-notes`, and fixed
the title of `extending-heckballs`.